requirements.txt
torch==2.1.2
torchvision==0.16.2
diffusers==0.27.2
transformers==4.37.2
accelerate==0.27.2
safetensors==0.4.2
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.1
python-multipart==0.0.6
scikit-learn==1.4.1
tqdm==4.66.1
tensorboard==2.15.1
albumentations==1.4.5
pyyaml==6.0.1
python-dotenv==1.0.0
matplotlib==3.8.3
